requirements.txt
diffusers==0.21.4
gpustat==1.1.1
gradio==3.47.1
py_cpuinfo==8.0.0
torch==2.0.1
transformers==4.34.0
